main/requirements.txt
certifi==2020.6.20
Django==4.0.4
djangorestframework==3.13.1
mkl-fft==1.3.1
mkl-service==2.4.0
pandas==1.4.1
pytz==2021.3
